another

+++ animrec/videoin.py
@@ -122,8 +122,8 @@ class VideoIn:
             fixx, fixy = alimu.fix_vidshape(self.res, self.maxres)
             print((rx1,ry1),(rx2,ry2))
             print(fixx, fixy)
-            fixx = ((self.maxres[0]/2.)-rx1)/(self.maxres[0]/2.)*fixx
-            fixy = ((self.maxres[1]/2.)-ry1)/(self.maxres[1]/2.)*fixy
+            fixx = int(((self.maxres[0]/2.)-rx1)/(self.maxres[0]/2.)*fixx)
+            fixy = int(((self.maxres[1]/2.)-ry1)/(self.maxres[1]/2.)*fixy)
             print(fixx, fixy)
             self.roil = ((rx1+fixx, ry1+fixy),(rx2, ry2))
             print(self.roil)
